Rigidbody density map generator WIP.

+++ VFD/Source/Core/Math/GaussQuadrature.h
@@ -6,7 +6,7 @@
 namespace vfd {
 	class GaussQuadrature {
 	public:
-		using Integrand = std::function<double(const glm::vec3&)>;
+		using Integrand = std::function<double(const glm::dvec3&)>;
 		
 		static double Integrate(const Integrand& integrand, const BoundingBox<glm::dvec3>& domain, unsigned int p);
 	};
